+++ R/api.R
@@ -53,6 +53,7 @@ vetiver_api <- function(pr,
                         ...) {
     # `force()` all `...` arguments early; https://github.com/tidymodels/vetiver/pull/20
     rlang::list2(...)
+    vetiver_model$model <- bundle::unbundle(vetiver_model$model)
     pr <- vetiver_pr_post(
         pr = pr,
         vetiver_model = vetiver_model,